Fixed the warning.
git-svn-id: svn://svn.rockbox.org/rockbox/trunk@7366 a1c6a512-1295-4272-9138-f99709370657

+++ apps/codecs/lib/xxx2wav.c
@@ -37,7 +37,7 @@ void* codec_malloc(size_t size)
 {
     void* x;
 
-    if (mem_ptr + size > bufsize)
+    if (mem_ptr + (int)size > bufsize)
         return NULL;
     
     x=&mallocbuf[mem_ptr];
